generating a playlist endpoint on backend and frontend is now modified.. new fields added to local playlist state and fb playlist database.. finally restructure of playlist module, broken up into playlist and playControl module
