+++ functions/api/controllers/playlistController.js
@@ -6,8 +6,8 @@ const spotifyWebApi = new SpotifyWebApi({
   clientSecret: keys.CLIENT_SECRET
 });
 
-exports.savePlaylist = (req, res, next) => {
-  let { access_token, playlistIds, playlistName } = req.body.data;
+exports.createPlaylist = (req, res, next) => {
+  let { access_token, playlistName } = req.body.data;
 
   spotifyWebApi.setAccessToken(access_token);
   spotifyWebApi
@@ -16,13 +16,47 @@ exports.savePlaylist = (req, res, next) => {
       return spotifyWebApi.createPlaylist(data.body.id, playlistName);
     })
     .then(data => {
-      return spotifyWebApi.addTracksToPlaylist(data.body.id, playlistIds);
+      res.status(200).json({
+        confirmation: 'success',
+        data: {
+          items: {
+            id: data.body.id
+          }
+        }
+      });
+      return;
     })
+    .catch(err => {
+      console.log(err);
+      if (err.statusCode === 401) {
+        res.status(401).json({
+          confirmation: 'fail',
+          message: err.message,
+          statusCode: 401
+        });
+      } else {
+        res.status(404).json({
+          confirmation: 'fail',
+          message: err.message,
+          statusCode: 404
+        });
+      }
+    });
+};
+
+exports.addToPlaylist = (req, res, next) => {
+  let { access_token, listOfPlaylistUris, playlistId } = req.body.data;
+
+  spotifyWebApi.setAccessToken(access_token);
+  spotifyWebApi
+    .addTracksToPlaylist(playlistId, listOfPlaylistUris)
     .then(data => {
       res.status(200).json({
         confirmation: 'success',
         data: {
-          items: data.body.snapshot_id
+          items: {
+            snapshot_id: data.body.snapshot_id
+          }
         }
       });
       return;
