+++ functions/api/routes/playlist.js
@@ -4,6 +4,7 @@ const router = express.Router();
 const PlaylistController = require('../controllers/playlistController');
 
 // routes
-router.post('/save', PlaylistController.savePlaylist);
+router.post('/create', PlaylistController.createPlaylist);
+router.post('/addtracks', PlaylistController.addToPlaylist);
 
 module.exports = router;
