+++ functions/client/src/store/index.js
@@ -2,6 +2,7 @@ import Vuex from 'vuex';
 import Vue from 'vue';
 import Playlist from './modules/Playlist';
 import Search from './modules/Search';
+import PlayerControl from './modules/PlayerControl';
 import Auth from './modules/Auth';
 Vue.use(Vuex);
 
@@ -9,6 +10,7 @@ export default new Vuex.Store({
   modules: {
     Playlist,
     Search,
-    Auth
+    Auth,
+    PlayerControl
   }
 });
