+++ functions/client/src/store/modules/PlayerControl/index.js
@@ -1,0 +1,11 @@
+import state from './state';
+import getters from './getters';
+import actions from './actions';
+import mutations from './mutations';
+
+export default {
+  state,
+  getters,
+  actions,
+  mutations
+};
