+++ functions/client/src/store/modules/Playlist/getters.js
@@ -1,23 +1,7 @@
 const getters = {
   getRecentlyGeneratedPlaylist: state => state.recentlyGeneratedPlaylist,
   getCurrentPlaylistMetaData: state => state.currentPlaylistMetaData,
-  getCurrentPlaylist: state => state.playlist,
-  getCurrentTrack: state => state.currentTrack,
-  getCurrentArtwork: state => state.currentArtwork,
-  getAudioElement: state => state.audioElement,
-  getShuffledPlaylist: state => state.shuffledPlaylist,
-  isPlaying: state => state.playing,
-  getCurrentTime: state => state.currentTime,
-  getProgress: state => state.progress,
-  getRemainingTime: state => state.remainingTime,
-  getVolume: state => state.volume,
-  isMouseDown: state => state.mouseDown,
-  getDuration: state => state.duration,
-  getCurrentTrackIndex: state => state.currentTrackIndex,
-  isAutoPlay: state => state.autoPlay,
-  isRepeat: state => state.repeat,
-  isMute: state => state.mute,
-  isShuffle: state => state.shuffle
+  getCurrentPlaylist: state => state.playlist
 };
 
 export default getters;
