+++ functions/client/src/store/modules/Playlist/state.js
@@ -1,23 +1,7 @@
 const state = {
   playlist: undefined,
   currentPlaylistMetaData: undefined,
-  recentlyGeneratedPlaylist: [],
-  currentTrack: undefined,
-  currentArtwork: '',
-  audioElement: undefined,
-  shuffledPlaylist: [],
-  playing: true,
-  currentTime: '0:00',
-  progress: 0,
-  remainingTime: '',
-  volume: 0,
-  mouseDown: false,
-  duration: 0,
-  currentTrackIndex: 0,
-  autoPlay: false,
-  repeat: false,
-  mute: false,
-  shuffle: false
+  recentlyGeneratedPlaylist: []
 };
 
 export default state;
